requirements.txt
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.6
lxml>=4.0.0
matplotlib==2.2.3
numpy==1.15.0
Pillow==5.2.0
pympi-ling==1.69
pydub==0.22.1
pathlib==1.0.1
scipy==1.1.0
sqlalchemy>=1.2.10
flask-googlelogin==0.3.1
flask-login==0.2.11
mysqlclient===1.3.13
